refactor: Input 로직을 Interface로 분리
